+++ packages/app/src/runner/SnapshotControls.spec.tsx
@@ -1,25 +1,36 @@
 import SnapshotControls from './SnapshotControls.vue'
 import { autSnapshot } from '../../cypress/support/fixtures'
 import { useSnapshotStore } from './snapshot-store'
+import { createEventManager, createTestAutIframe } from '../../cypress/e2e/support/ctSupport'
+
+// function createTestAutIframe () {
+//   return new class {
+//     removeHighlights () {}
+//   }
+// }
 
 describe('SnapshotControls', () => {
-  const mountSnapshotControls = () => {
-    const eventManager = {} as typeof window.UnifiedRunner.eventManager
-
-    return cy.mount(() => <SnapshotControls eventManager={eventManager} />)
+  const mountSnapshotControls = (
+    eventManager = createEventManager(),
+    autIframe = createTestAutIframe(),
+  ) => {
+    return cy.mount(() => (
+      <SnapshotControls
+        eventManager={eventManager}
+        getAutIframe={() => autIframe}
+      />
+    ))
   }
 
-  beforeEach(() => {
-    return mountSnapshotControls()
-  })
-
   it('renders nothing when messageTitle is undefined', () => {
+    mountSnapshotControls()
     cy.get('[data-cy="snapshot-highlight-controls"]').should('not.exist')
     cy.get('[data-cy="snapshot-message"]').should('not.exist')
     cy.get('[data-cy="snapshot-change-state"]').should('not.exist')
   })
 
   it('renders snapshot title when one is pinned', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.pinSnapshot(autSnapshot)
@@ -28,6 +39,7 @@ describe('SnapshotControls', () => {
   })
 
   it('renders snapshot pinned status', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.pinSnapshot(autSnapshot)
@@ -40,6 +52,7 @@ describe('SnapshotControls', () => {
   })
 
   it('clears snapshot message', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.pinSnapshot(autSnapshot)
@@ -49,6 +62,7 @@ describe('SnapshotControls', () => {
   })
 
   it('shows snapshot with custom message', () => {
+    mountSnapshotControls()
     const message = 'This is a custom message'
     const snapshotStore = useSnapshotStore()
 
@@ -57,6 +71,7 @@ describe('SnapshotControls', () => {
   })
 
   it('does not show highlight controls if no element present on snapshot', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.pinSnapshot(autSnapshot)
@@ -65,12 +80,30 @@ describe('SnapshotControls', () => {
 
   it('toggles highlight controls if snapshot has an element', () => {
     const snapshotStore = useSnapshotStore()
+    const eventManager = createEventManager()
+    const autIframe = createTestAutIframe()
+    const removeHighlights = cy.stub(autIframe, 'removeHighlights')
 
+    // we don't have an iframe-model since this is a CT test, but we can
+    // simulate it by registering the same unpin:snapshot event it does.
+    eventManager.on('unpin:snapshot', () => snapshotStore.unpinSnapshot())
     snapshotStore.pinSnapshot({ ...autSnapshot, $el: 'some element' })
+
+    mountSnapshotControls(eventManager, autIframe)
     cy.get('[data-cy="snapshot-highlight-controls"]').should('exist')
+    cy.get('[data-cy="toggle-snapshot-highlights"]').as('toggle')
+    cy.get('@toggle').should('have.attr', 'title', 'Hide highlights')
+    cy.get('@toggle').click().then(() => {
+      expect(removeHighlights).to.have.been.calledOnce
+    })
+
+    cy.get('@toggle').should('have.attr', 'title', 'Show highlights')
+    cy.get('[data-cy="unpin-snapshot"]').click()
+    cy.get('[data-cy="snapshot-highlight-controls"]').should('not.exist')
   })
 
   it('shows running test error', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.setTestsRunningError()
@@ -78,6 +111,7 @@ describe('SnapshotControls', () => {
   })
 
   it('shows snapshot missing error', () => {
+    mountSnapshotControls()
     const snapshotStore = useSnapshotStore()
 
     snapshotStore.setMissingSnapshotMessage()